A new user type, developer who can see only their own reports and databases
git-svn-id: http://192.168.1.10/repos/toto@2077 b53939d4-f08c-4181-96d4-b201aa9c61c4

+++ Toto/trunk/src/com/azquo/admin/controller/ManageReportsController.java
@@ -62,30 +62,28 @@ public class ManageReportsController {
             , @RequestParam(value = "deleteId", required = false) String deleteId
             , @RequestParam(value = "databaseIdList", required = false) String[] databaseIdList // I think this is correct?
             , @RequestParam(value = "name", required = false) String name
-            , @RequestParam(value = "category", required = false) String category
             , @RequestParam(value = "explanation", required = false) String explanation
             , @RequestParam(value = "submit", required = false) String submit
     ) {
         LoggedInUser loggedInUser = (LoggedInUser) request.getSession().getAttribute(LoginController.LOGGED_IN_USER_SESSION);
         // I assume secure until we move to proper spring security
-        if (loggedInUser == null || !loggedInUser.getUser().isAdministrator()) {
-            return "redirect:/api/Login";
-        } else {
-            if (deleteId != null){
-                AdminService.removeReportById(loggedInUser,Integer.parseInt(deleteId));
+        if (loggedInUser != null && (loggedInUser.getUser().isAdministrator() || loggedInUser.getUser().isDeveloper())) {
+            if (deleteId != null) {
+                AdminService.removeReportById(loggedInUser, Integer.parseInt(deleteId));
             }
             if (editId != null) {
                 final OnlineReport theReport = OnlineReportDAO.findById(Integer.parseInt(editId)); // yes could exception, don't care at the mo
-                if (theReport.getBusinessId() == loggedInUser.getUser().getBusinessId()) {
+                if (theReport != null && ((loggedInUser.getUser().isAdministrator() && theReport.getBusinessId() == loggedInUser.getUser().getBusinessId())
+                        || (loggedInUser.getUser().isDeveloper() && theReport.getUserId() == loggedInUser.getUser().getId()))) {
                     // ok check to see if data was submitted
                     if (submit != null) {
                         // to keep intelliJ happy, I'm not sure if this is a good idea or not? I suppose protects against logic above being changed unpredictably
                         DatabaseReportLinkDAO.unLinkReport(theReport.getId());
-                        for (String databaseId : databaseIdList){
+                        // todo - stop db id hacking here by developers
+                        for (String databaseId : databaseIdList) {
                             DatabaseReportLinkDAO.link(Integer.parseInt(databaseId), theReport.getId());
                         }
                         theReport.setReportName(name);
-                        theReport.setReportCategory(category);
                         theReport.setExplanation(explanation);
                         OnlineReportDAO.store(theReport);
                         model.put("reports", AdminService.getReportList(loggedInUser));
@@ -94,21 +92,24 @@ public class ManageReportsController {
                     final List<Integer> databaseIdsForReportId = DatabaseReportLinkDAO.getDatabaseIdsForReportId(theReport.getId());
                     model.put("id", editId);
                     List<DatabaseSelected> databasesSelected = new ArrayList<>();
-                    final List<Database> forBusinessId = DatabaseDAO.findForBusinessId(loggedInUser.getUser().getBusinessId());
-                    for (Database database : forBusinessId){
+
+                    final List<Database> forUser = AdminService.getDatabaseListForBusiness(loggedInUser);
+                    for (Database database : forUser) {
                         databasesSelected.add(new DatabaseSelected(databaseIdsForReportId.contains(database.getId()), database));
                     }
                     model.put("databasesSelected", databasesSelected);
                     model.put("name", theReport.getReportName());
                     model.put("file", theReport.getFilename());
-                    model.put("category", theReport.getReportCategory() != null ? theReport.getReportCategory() : "");
                     model.put("explanation", theReport.getExplanation() != null ? theReport.getExplanation() : "");
                     return "editreport";
                 }
             }
             // if not editing then very simple
             model.put("reports", AdminService.getReportList(loggedInUser));
+            model.put("developer", loggedInUser.getUser().isDeveloper());
             return "managereports";
+        } else {
+            return "redirect:/api/Login";
         }
     }
 }
